requirements.txt
arrow
csirtg_indicator>=3.0a1,<4.0
dnspython
csirtg-dnsdb==0.0.0a4